waldobook/pyproject.toml
[tool.poetry]
name = "waldobook"
version = "0.1.0"
description = ""
authors = ["Max Katz-Christy <maxkatzchristy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
SQLAlchemy = "^1.4.46"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
qrcode = {extras = ["pil"], version = "^7.3.1"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
